pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-code"
version = "0.1.0"
description = "CLI tool for interacting with Gemini 2.5 Pro"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "click>=8.0.0",
    "rich>=10.0.0",
    "google-generativeai>=0.3.0",
    "pyyaml>=6.0",
    "tiktoken>=0.4.0",
    "requests>=2.25.0",
    "html2text>=2020.1.16",
]

[project.scripts]
gemini = "gemini_cli.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}